Fix Symfony 5.3 compatibility
