+++ DependencyInjection/Security/Factory/JWTAuthenticatorFactory.php
@@ -3,74 +3,51 @@
 namespace Lexik\Bundle\JWTAuthenticationBundle\DependencyInjection\Security\Factory;
 
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\AuthenticatorFactoryInterface;
-use Symfony\Component\Config\Definition\Builder\NodeDefinition;
-use Symfony\Component\DependencyInjection\ChildDefinition;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 
-/**
- * Wires the "jwt" authenticator from user configuration.
- *
- * @author Robin Chalas <robin.chalas@gmail.com>
- */
-class JWTAuthenticatorFactory implements AuthenticatorFactoryInterface
-{
-    /**
-     * @throws \LogicException
-     */
-    public function create(ContainerBuilder $container, $id, $config, $userProvider, $defaultEntryPoint)
-    {
-        throw new \LogicException('This method is implemented for BC purpose and should never be called.');
-    }
+if (interface_exists(SecurityFactoryInterface::class) && !interface_exists(AuthenticatorFactoryInterface::class)) {
+    eval('
+        namespace Lexik\Bundle\JWTAuthenticationBundle\DependencyInjection\Security\Factory;
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getPriority(): int
-    {
-        return -10;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getKey(): string
-    {
-        return 'jwt';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addConfiguration(NodeDefinition $node)
-    {
-        $node
-            ->children()
-                ->scalarNode('provider')
-                    ->defaultNull()
-                ->end()
-                ->scalarNode('authenticator')
-                    ->defaultValue('lexik_jwt_authentication.security.jwt_authenticator')
-                ->end()
-            ->end()
-        ;
-    }
-
-    public function createAuthenticator(ContainerBuilder $container, string $firewallName, array $config, string $userProviderId): string
-    {
-        $authenticatorId = 'security.authenticator.jwt.'.$firewallName;
+        use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\AuthenticatorFactoryInterface;
+        use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
         
-        $userProviderId = empty($config['provider']) ? $userProviderId : 'security.user.provider.concrete.' . $config['provider'];
+        /**
+         * Wires the "jwt" authenticator from user configuration.
+         *
+         * @author Robin Chalas <robin.chalas@gmail.com>
+         */
+        class JWTAuthenticatorFactory implements SecurityFactoryInterface
+        {
+            use JWTAuthenticatorFactoryTrait;
+        }
+    ');
+} elseif (!method_exists(SecurityExtension::class, 'addAuthenticatorFactory')) {
+    eval('
+        namespace Lexik\Bundle\JWTAuthenticationBundle\DependencyInjection\Security\Factory;
 
-        $container
-            ->setDefinition($authenticatorId, new ChildDefinition($config['authenticator']))
-            ->replaceArgument(3, new Reference($userProviderId))
-        ;
-
-        // Compile-time parameter removed by RemoveLegacyAuthenticatorPass
-        // Stop setting it when guard support gets removed (aka when removing Symfony<5.3 support)
-        $container->setParameter('lexik_jwt_authentication.authenticator_manager_enabled', true);
-
-        return $authenticatorId;
+        use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\AuthenticatorFactoryInterface;
+        use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
+        
+        /**
+         * Wires the "jwt" authenticator from user configuration.
+         *
+         * @author Robin Chalas <robin.chalas@gmail.com>
+         */
+        class JWTAuthenticatorFactory implements AuthenticatorFactoryInterface, SecurityFactoryInterface
+        {
+            use JWTAuthenticatorFactoryTrait;
+        }
+    ');
+} else {
+    /**
+     * Wires the "jwt" authenticator from user configuration.
+     *
+     * @author Robin Chalas <robin.chalas@gmail.com>
+     */
+    class JWTAuthenticatorFactory implements AuthenticatorFactoryInterface
+    {
+        use JWTAuthenticatorFactoryTrait;
     }
 }
