+++ composer.json
@@ -41,27 +41,27 @@
         "ext-openssl": "*",
         "lcobucci/jwt": "^3.4|^4.0",
         "namshi/jose": "^7.2",
-        "symfony/config": "^4.4|^5.4|^6.0",
-        "symfony/dependency-injection": "^4.4|^5.4|^6.0",
+        "symfony/config": "^4.4|^5.3|^6.0",
+        "symfony/dependency-injection": "^4.4|^5.3|^6.0",
         "symfony/deprecation-contracts": "^2.4",
-        "symfony/event-dispatcher": "^4.4|^5.4|^6.0",
-        "symfony/http-foundation": "^4.4|^5.4|^6.0",
-        "symfony/http-kernel": "^4.4|^5.4|^6.0",
-        "symfony/property-access": "^4.4|^5.4|^6.0",
-        "symfony/security-bundle": "^4.4|^5.4|^6.0",
-        "symfony/security-core": "^4.4|^5.4|^6.0",
-        "symfony/security-http": "^4.4|^5.4|^6.0"
+        "symfony/event-dispatcher": "^4.4|^5.3|^6.0",
+        "symfony/http-foundation": "^4.4|^5.3|^6.0",
+        "symfony/http-kernel": "^4.4|^5.3|^6.0",
+        "symfony/property-access": "^4.4|^5.3|^6.0",
+        "symfony/security-bundle": "^4.4|^5.3|^6.0",
+        "symfony/security-core": "^4.4|^5.3|^6.0",
+        "symfony/security-http": "^4.4|^5.3|^6.0"
     },
     "require-dev": {
-        "symfony/browser-kit": "^4.4|^5.4|^6.0",
-        "symfony/console": "^4.4|^5.4|^6.0",
-        "symfony/dom-crawler": "^4.4|^5.4|^6.0",
-        "symfony/filesystem": "^4.4|^5.4|^6.0",
-        "symfony/framework-bundle": "^4.4|^5.4|^6.0",
-        "symfony/phpunit-bridge": "^4.4|^5.4|^6.0",
-        "symfony/security-guard": "^4.4|^5.4",
-        "symfony/var-dumper": "^4.4|^5.4|^6.0",
-        "symfony/yaml": "^4.4|^5.4|^6.0"
+        "symfony/browser-kit": "^4.4|^5.3|^6.0",
+        "symfony/console": "^4.4|^5.3|^6.0",
+        "symfony/dom-crawler": "^4.4|^5.3|^6.0",
+        "symfony/filesystem": "^4.4|^5.3|^6.0",
+        "symfony/framework-bundle": "^4.4|^5.3|^6.0",
+        "symfony/phpunit-bridge": "^4.4|^5.3|^6.0",
+        "symfony/security-guard": "^4.4|^5.3",
+        "symfony/var-dumper": "^4.4|^5.3|^6.0",
+        "symfony/yaml": "^4.4|^5.3|^6.0"
     },
     "conflict": {
         "symfony/console": "<4.4"
